requirements/constraints.txt
# Version constraints for pip-installation.
#
# This file doesn't install any packages. It specifies version constraints
# that will be applied if a package is needed.
#
# When pinning something here, please provide an explanation of why.  Ideally,
# link to other information that will help people in the future to remove the
# pin when possible.  Writing an issue against the offending project and
# linking to it here is good.

# openedx Mango version is working with 3.2.
Django<4.0
# json api has to be compatible with rest-framework 3.12.4(maple)
djangorestframework-jsonapi==5.0.0
djangorestframework==3.12.4
# maple django edx rest extensions
edx-drf-extensions==8.0.0
# maple django filter version
django-filter==21.1
